feat: Optimize getRequestLogsForGraph query
The getRequestLogsForGraph function in the api-controller.ts file has been optimized by using a query object to filter the logs. This improves the performance of retrieving request logs for a specific app, server, request type, and timestamp range. Additionally, the query now includes an optional filter for the apiId. This change enhances the efficiency and flexibility of the getRequestLogsForGraph functionality.

Co-authored-by: Kshitij Ayush <kshitij4ka@gmail.com>